Add helper method for responding to input
Add new internal method that figures out if the response shold go back
to the same target as the event being responded to or if instead the
response should go to the user that sent that message. The choise is
based on whether or not the event target is the current bot nick (i.e.
a PM).