Require requests[security] for SNI on Python > 2.7.9
The Requests library will automatically use PyOpenSSL to provide SNI
(server name indication) support for older versions of Python where
urllib3 does not natively handle it. This is useful when talking to
Phabricator and other web services that may be behind a unified TLS
gateway which requires SNI indications during the TLS handshake.