Log channel from on_topic
Prior use of on_join was wrong. JOIN events are sent for each user that
joins the channel. The server should send TOPIC notifications when our
client first joins the channel. The bot will get notified when the
channel topic is changed too, but that is a lot less common than new
users attaching.