Use messages to trigger authentication and channel joins
Listen for privmsgs from NickServ to determine if auth is needed and if
so when it succeeds or fails. When a password is provided in the
configuration do not join channels until NickServ tells us that
authentication worked. When joining channels, use execute_delayed rather
than sleep to space out joins.

Inspired by fidibot by @nickraptis:
https://github.com/nickraptis/fidibot/blob/61d97a7/src/fidibot.py