Honor ACLs for !log processing (#17)
Allow configuring whitelist/blacklist patterns for controlling which
!log messages will actually be processed. The access control list is
configurable per-channel. ACLs are expressed as ban mask style glob
patterns.

If the !log message source is not allowed based on the ACL configuration
a warning message is logged and the user is told that they are not
authorized. The logging should help fine tune configuration of the ACLs.