profile: make less output nicer
I *hate* the default behavior of less to clear the screen on exit.
I almost always want to search around to find something interesting in
the file and then exit out to do something with that data. I don't want
to have to remember to cut-n-paste things before closing it.

Change-Id: I59f65c0b6cb82acd6d4947ab773aadb04cd14d33