Use IB3 library for basic functionality
Use the IB3 library (<https://github.com/bd808/python-ib3/>) to provide
basic bot functionality:
* SASL auth
* SSL connections to Freenode
* Disconnect from Freenode on error
* Ping server to check for connectivity
* Rejoin channels when banned
* Rejoin channels when kicked
* Regain primary nick when taken
* Use lenient UTF-8 decoding for messages

Much of the IB3 library was extracted from Stashbot itself, but the SASL
auth and SSL connection support are new functionality.

When deployed, the virtual environment will need to be updated to
install the new library from PyPi. The port for connecting to Freenode
should also be changed to 6697 which is the recommended canonical port
for TLS/SSL encrypted IRC connections.

Change-Id: I380672e4ee0e65f5bd12f6e8d0748aae46cd0635