Only ignore phab mentions
wikibugs has a feature where the deployment script sends a !log message
to #wikimedia-cloud. Unfortunately wikibugs is on the active ignore list
so that channels are not spammed with Phabricator announcements as
a result of talk by other bots. Change the use of the ignore filtering
to take effect after !log messages are processed so that wikibugs can be
cool and log its code changes.

Change-Id: If5ac579bcc02b060b9f79cdef25da33b97b963b3