!log: only update Phabricator tasks if we are responding in channel
Add a guard condition to stashbot.sal._store_in_es() to prevent
Phabricator task comments from being created. Set this guard condition
to the same value as stashbot.sal.log()'s respond_to_channel flag when
processing log events.

There is a funky edge case for !log messages in #wikimedia-cloud
and #wikimedia-operations which are related to the Release Engineering
team. These messages are written both to the main SAL for the channel
the message originates in and to the Release Engineering SAL. An
unintended side effect of this double logging was that any Phabricator
tasks mentioned in the log message would get two Stashbot mentions (one
for the original SAL and a second for the RelEng SAL).

Change-Id: I7c078e538440b186f8ce9cf17e5f85cb824859e4