Fix nested dict iteration for py3
In py2, dict.keys() gives you a list of the keys in the dict
at the time of the call. In py3, this was optimized to return an
iterable that is backed by the life dict. The py3 behavior has a better
memory footprint, but it introduces the possibility of "RuntimeError:
dictionary changed size during iteration" failures if the dict is
mutated during iteration. Using `list(dict.keys())` in py3 gives
a behavior similar to the py2 default.

Change-Id: I57b9d54651656af5d225fe28265f80e90196fed9