Document how to find available message placeholders
Also update the example to feature the project, which I'm proposing
we use in the deployed config. It seems the example was already a bit
outdated as I'm seeing `channel` used in Phab, but not yet reflected
here.

I considered putting both in there, but that would mean logging
`#wikimedia-operations production` or `#wikimedia-releng releng`,
which seems needlessly verbose.

If we do want to log the channel in some cases, we could alternatively
modify Logger.log() and introduce a new `bang` property that is e.g.
a pretty-printed string that holds the channel for simple cases where
there is only one project for that channel, and `<channel> <project>`
for #wikimedia-cloud where there are multiple, basically reflecting
how the `!log` messages are on IRC. But maybe the project suffices?

Bug: T186845
Change-Id: I87284d4851465c9b9c596fa66113d3f9372f25e7