sal: Fix detection of #releng in #wikimedia-operations !log
In 5ecbfa5 (~5 years ago) we introduced special support for collecting
!log events related to the Wikimedia Foundation's Release Engineering
team into a dedicated SAL. This included support for "tagging" !log
events sent to the #wikimedia-operations IRC channel with `#releng` to
duplicate the logged event to the Release Engineering SAL. Nobody
noticed however that the guard condition for handling this case was over
indented! As written the check to see if the event came from the
-operations channel was nested inside an existing branch for messages
sent to the #wikimedia-cloud channel.

Fix by making the `elif channel == "#wikimedia-operations"` conditional
a sibling of the `if channel in ["#wikimedia-labs", "#wikimedia-cloud"]`
conditional rather than a child.

Change-Id: I41bd8f2c67fd6bd2adbf3706ea4c3cf6d6158054