Sync etc/ and bin/ with changes from deployment
Sync changes made in the live deployment while migrating from Fleanode
to Libera.Chat with the repo. We could really roll all of these back,
but it seems nice to leave a pattern that could be reused in the future
if we find a need to migrate to another IRC network or a completely
different chatops gateway.

Change-Id: Ia4ed511886e8dca3748ea64edf4533613d60a396