Parse user@host in wm-bot2 messages
After the change I4bdae1182ef3d5862a6dc351ba3112a29d3d27d6,
wmcs-cookbooks are prepending user@host to all messages sent via wm-bot.

We can now parse those messages as we do for messages by logmsgbot, but
adding a "wmbot~" prefix to prevent people from spoofing it (as wm-bot
is open to anyone on the internet so we cannot trust that user@host is
real and not spoofed).

Bug: T346631
Change-Id: Ic0c6a5410c21921081810a21b868a5406c1c703e